scraper/requirements.txt
# Web Scraping
requests>=2.28.0
beautifulsoup4>=4.11.0
httpx>=0.24.0
fake-useragent>=1.1.0
lxml>=4.9.0

# AI / Embeddings
sentence-transformers>=2.2.0
numpy>=1.24.0
scikit-learn>=1.2.0

# Database
supabase>=2.0.0
python-dotenv>=1.0.0

# Utils
pandas>=2.0.0
tqdm>=4.65.0
schedule>=1.2.0
